requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.2.0
ansicon==1.89.0
anyio==4.2.0
arrow==1.2.3
asttokens==2.0.5
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.6.2
boto3==1.34.82
botocore==1.34.82
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==4.2.4
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
croniter==1.3.7
cryptography==42.0.5
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.7.1
einops==0.8.0
executing==0.8.3
fastapi==0.103.0
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.0
fsspec==2024.3.1
gmpy2==2.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.1
grpcio==1.48.2
h11==0.14.0
idna==3.7
inquirer==3.1.4
ipykernel==6.28.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.4
jinxed==1.2.0
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
lightning==2.0.9.post0
lightning-cloud==0.5.57
lightning-utilities==0.9.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.4
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.3.0
pip==24.0
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.3
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
pyzmq==25.1.2
readchar==4.0.5
requests==2.32.2
requests-oauthlib==1.3.0
rich==13.3.5
rsa==4.7.2
s3transfer==0.10.1
seaborn==0.13.2
setuptools==69.5.1
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0
tensorboardX==2.2
torch==2.2.2
torch-tb-profiler==0.4.3
torchaudio==2.2.2
torchmetrics==1.1.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.4
traitlets==5.7.1
typing_extensions==4.11.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.2.1
uvicorn==0.20.0
wcwidth==0.2.5
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
yarl==1.9.3
